+++ examples/upload.hs
@@ -4,6 +4,7 @@ module Main (main) where
 import Web.Scotty
 
 import Control.Monad.IO.Class
+import qualified Data.Text.Lazy as TL
 
 import Network.Wai.Middleware.RequestLogger
 import Network.Wai.Middleware.Static
@@ -16,8 +17,6 @@ import Text.Blaze.Html.Renderer.Text (renderHtml)
 import qualified Data.ByteString.Lazy as B
 import qualified Data.ByteString.Char8 as BS
 import System.FilePath ((</>))
-import Prelude ()
-import Prelude.Compat
 
 main :: IO ()
 main = scotty 3000 $ do
@@ -41,7 +40,7 @@ main = scotty 3000 $ do
         -- write the files to disk, so they will be served by the static middleware
         liftIO $ sequence_ [ B.writeFile ("uploads" </> fn) fc | (_,fn,fc) <- fs' ]
         -- generate list of links to the files just uploaded
-        html $ mconcat [ mconcat [ fName
+        html $ mconcat [ mconcat [ TL.fromStrict fName
                                  , ": "
                                  , renderHtml $ H.a (H.toHtml fn) H.! (href $ H.toValue fn) >> H.br
                                  ]